+++ server/src/main/java/org/opensearch/index/IndexSettings.java
@@ -1063,11 +1063,11 @@ public final class IndexSettings {
     }
 
     public boolean isSegRepLocalEnabled() {
-        return isSegRepEnabled() && !isSegRepWithRemoteEnabled();
+        return isSegRepEnabled() && !isRemoteStoreEnabled();
     }
 
     public boolean isSegRepWithRemoteEnabled() {
-        return isSegRepEnabled() && isRemoteStoreEnabled() && FeatureFlags.isEnabled(FeatureFlags.SEGMENT_REPLICATION_EXPERIMENTAL);
+        return isSegRepEnabled() && isRemoteStoreEnabled();
     }
 
     /**